Fix accessing https url permission, auth err msg.
Jersey's client API requires additional permissions when it tries to
access https urls, it looks like.

When authentication against VideoScribe's API fails, the UI displays an
empty error message due to a missing case. Added a generic error message
to be shown on the UI side if we encounter an uncategorized 400 error.